fix(dev_auth) improving auth controller, exception handler and moving activation method
