+++ auth/src/main/java/com/epam/esm/auth/controller/AuthenticationController.java
@@ -4,10 +4,9 @@ import com.epam.esm.auth.models.AuthenticationRequest;
 import com.epam.esm.auth.models.RegisterRequest;
 import com.epam.esm.auth.models.TokenDTO;
 import com.epam.esm.auth.service.AuthenticationService;
+import com.epam.esm.credentials.service.CredentialsService;
 import com.epam.esm.jwt.service.JwtService;
 import com.epam.esm.model.UserDTO;
-import jakarta.annotation.Nullable;
-import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -24,11 +23,12 @@ import static com.epam.esm.utils.Constants.AUTHORIZATION_HEADER;
 @RequiredArgsConstructor
 public class AuthenticationController {
     private final AuthenticationService authenticationService;
+    private final CredentialsService credentialsService;
     private final JwtService jwtService;
 
     @PostMapping(path = "/register", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
     public ResponseEntity<String> register(@Valid @RequestPart("request") RegisterRequest request,
-                                           @RequestPart("image") @Nullable MultipartFile image) {
+                                           @RequestPart(value = "image", required = false) MultipartFile image) {
         return ResponseEntity.ok(authenticationService.register(request, image));
     }
 
@@ -38,8 +38,8 @@ public class AuthenticationController {
     }
 
     @PostMapping("/refresh-token")
-    public ResponseEntity<Void> refreshToken(HttpServletRequest request, HttpServletResponse response) {
-        String bearerToken = request.getHeader(AUTHORIZATION_HEADER);
+    public ResponseEntity<Void> refreshToken(@RequestHeader(AUTHORIZATION_HEADER) String bearerToken,
+                                             HttpServletResponse response) {
         jwtService.validateToken(bearerToken);
         String newAccessToken = authenticationService.refreshToken(bearerToken.substring(7));
         response.setHeader(AUTHORIZATION_HEADER, AUTHENTICATION_BEARER_TOKEN + newAccessToken);
@@ -48,7 +48,7 @@ public class AuthenticationController {
 
     @PostMapping("/activate-account")
     public ResponseEntity<Void> activateAccount(@RequestParam String token) {
-        authenticationService.activateAccount(token);
+        credentialsService.activateAccount(token);
         return ResponseEntity.noContent().build();
     }
 
@@ -59,7 +59,7 @@ public class AuthenticationController {
     }
 
     @GetMapping("/user")
-    public ResponseEntity<UserDTO> getUserFromJwt(@RequestHeader("Authorization") String bearerToken) {
+    public ResponseEntity<UserDTO> getUserFromJwt(@RequestHeader(AUTHORIZATION_HEADER) String bearerToken) {
         jwtService.validateToken(bearerToken);
         return ResponseEntity.ok(authenticationService.decodeUserFromJwt(bearerToken.substring(7)));
     }
