+++ auth/src/main/java/com/epam/esm/auth/service/AuthenticationService.java
@@ -14,6 +14,4 @@ public interface AuthenticationService {
     String refreshToken(String jwt);
 
     UserDTO decodeUserFromJwt(String jwt);
-
-    void activateAccount(String token);
 }
