+++ auth/src/main/java/com/epam/esm/credentials/service/CredentialsService.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Service;
 
 @Service
 public interface CredentialsService {
-    void createCredentials(RegisterRequest request);
+    void create(RegisterRequest request);
 
     void activateAccount(String email);
