+++ auth/src/main/java/com/epam/esm/credentials/service/CredentialsServiceImpl.java
@@ -3,24 +3,34 @@ package com.epam.esm.credentials.service;
 import com.epam.esm.auth.models.RegisterRequest;
 import com.epam.esm.credentials.model.Credentials;
 import com.epam.esm.credentials.repository.CredentialsRepository;
+import com.epam.esm.jwt.TokenType;
+import com.epam.esm.jwt.service.JwtService;
+import com.epam.esm.utils.amqp.MessagePublisher;
 import com.epam.esm.utils.exceptionhandler.exceptions.AccountIsActiveException;
 import com.epam.esm.utils.exceptionhandler.exceptions.EmailAlreadyRegisteredException;
 import com.epam.esm.utils.exceptionhandler.exceptions.EmailNotFoundException;
+import com.epam.esm.utils.exceptionhandler.exceptions.InvalidTokenException;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-import static com.epam.esm.utils.Constants.EMAIL_IS_ALREADY_REGISTERED;
-import static com.epam.esm.utils.Constants.USER_NOT_EXIST_EMAIL;
+import static com.epam.esm.model.Role.ADMIN;
+import static com.epam.esm.utils.Constants.*;
 
 @Service
 @RequiredArgsConstructor
 public class CredentialsServiceImpl implements CredentialsService {
     private final CredentialsRepository credentialsRepository;
     private final PasswordEncoder passwordEncoder;
+    private final JwtService jwtService;
+    private final MessagePublisher messagePublisher;
+
 
     @Override
-    public void createCredentials(RegisterRequest request) {
+    public void create(RegisterRequest request) {
         credentialsRepository.findByEmail(request.email()).ifPresent(req -> {
             throw new EmailAlreadyRegisteredException(EMAIL_IS_ALREADY_REGISTERED);
         });
@@ -32,7 +42,12 @@ public class CredentialsServiceImpl implements CredentialsService {
     }
 
     @Override
-    public void activateAccount(String email) {
+    @Transactional
+    public void activateAccount(String token) {
+        String email = jwtService.extractUsername(token);
+        if (jwtService.extractType(token) != TokenType.EMAIL_VALIDATION || email.isEmpty()) {
+            throw new InvalidTokenException(INVALID_VALIDATION_TOKEN);
+        }
         Credentials credentials =
                 credentialsRepository.findByEmail(email).orElseThrow(() -> new EmailNotFoundException(USER_NOT_EXIST_EMAIL));
         if (credentials.isEnabled()) throw new AccountIsActiveException();
